rename: cpf_generator -> cpf_tools
